Player Header and Class
1.Created Player Header File--->added headers[pokemonType ,pokemonChoice ,Utility]--->class Player--->declaration ,data members and methods of player
2.Created Player Class file --->methods defined
